fix user overview bug;

+++ app/src/main/java/com/blue_stingray/healthy_life_app/model/AppGoal.java
@@ -3,8 +3,8 @@ package com.blue_stingray.healthy_life_app.model;
 public class AppGoal {
 
     public String name;
-    public int remaining;
-    public int spent;
-    public int total;
+    public double remaining;
+    public double spent;
+    public double total;
 
 }
